move removeDuplicateNestedTerms into its own file
