+++ packages/react-tei/src/unitex/enrichDocumentWithUnitex.spec.ts
@@ -6,92 +6,13 @@ import {
 	getRemainingStringParts,
 	getTermOverlap,
 	getWordOverlap,
-	hasIdenticalTermInSubTerms,
 	isContainedIn,
 	mergeIdenticalTerms,
 	nestContainedTerms,
-	removeDuplicateNestedTerms,
 	splitTermIntoSegments,
 } from "./enrichDocumentWithUnitex";
 
 describe("enrichDocumentWithUnitex", () => {
-	describe("hasIdenticalTermInSubTerms", () => {
-		it("should return true when a term has an identical term in a list of subTerms", () => {
-			const term = {
-				term: "America",
-				groups: ["group2"],
-			};
-
-			const subTerms = [
-				{
-					term: "States of America",
-					groups: ["group3"],
-				},
-				{
-					term: "America",
-					groups: ["group2"],
-				},
-			];
-			expect(hasIdenticalTermInSubTerms(term, subTerms)).toBe(true);
-		});
-
-		it("should return false when a term does not have an identical term in a list of subTerms", () => {
-			const term = {
-				term: "America",
-				groups: ["group2"],
-			};
-			const subTerms = [
-				{
-					term: "States of America",
-					groups: ["group3"],
-				},
-				{
-					term: "United States",
-					groups: ["group1"],
-				},
-			];
-			expect(hasIdenticalTermInSubTerms(term, subTerms)).toBe(false);
-		});
-
-		it("should return true when a term has an identical term in deeply nested list of subTerms", () => {
-			const term = {
-				term: " sec",
-				groups: ["group2"],
-			};
-			const subTerms = [
-				{
-					term: "Union ratatinée des saucissons sec",
-					groups: ["group3"],
-					subTerms: [
-						{
-							term: "Union ratatinée",
-							groups: ["group2"],
-						},
-						{
-							term: " des ",
-							groups: [],
-						},
-						{
-							term: "saucissons sec",
-							groups: ["group1"],
-							subTerms: [
-								{
-									term: "saucissons",
-									groups: ["group4"],
-								},
-								{
-									term: " sec",
-									groups: [],
-								},
-							],
-						},
-					],
-				},
-			];
-			expect(hasIdenticalTermInSubTerms(term, subTerms)).toBe(true);
-		});
-	});
-
 	describe("getRemainingStringParts", () => {
 		it("should return the remaining parts of a string after removing a term at the start", () => {
 			const terms = ["New York", "great"];
@@ -556,53 +477,6 @@ describe("enrichDocumentWithUnitex", () => {
 			const terms: { term: string; group: string }[] = [];
 			const mergedTerms = mergeIdenticalTerms(terms);
 			expect(mergedTerms).toEqual([]);
-		});
-	});
-
-	describe("removeDuplicateNestedTerms", () => {
-		it("should remove duplicate nested terms that are identical to one of the descendant of their brother", () => {
-			const terms = [
-				{
-					term: "United States of America",
-					groups: ["group1"],
-					subTerms: [
-						{
-							term: "America",
-							groups: ["group1", "group2"],
-						},
-						{
-							term: "States of America",
-							groups: ["group1", "group3"],
-							subTerms: [
-								{
-									term: "America",
-									groups: ["group1", "group3", "group2"],
-								},
-							],
-						},
-					],
-				},
-			];
-
-			const cleanedTerms = removeDuplicateNestedTerms(terms);
-			expect(cleanedTerms).toEqual([
-				{
-					term: "United States of America",
-					groups: ["group1"],
-					subTerms: [
-						{
-							term: "States of America",
-							groups: ["group1", "group3"],
-							subTerms: [
-								{
-									term: "America",
-									groups: ["group1", "group3", "group2"],
-								},
-							],
-						},
-					],
-				},
-			]);
 		});
 	});
